+++ leptos_hotkeys/src/lib.rs
@@ -1,8 +1,10 @@
 mod context;
+mod hotkey;
 mod macros;
 mod types;
 mod use_hotkeys;
 
 pub use context::{provide_hotkeys_context, use_hotkeys_context, HotkeysContext};
-pub use types::{Hotkey, KeyboardModifiers};
-pub use use_hotkeys::{use_hotkeys_ref_scoped, use_hotkeys_scoped};
+pub use hotkey::Hotkey;
+pub use types::KeyboardModifiers;
+pub use use_hotkeys::{use_hotkeys_ref, use_hotkeys_scoped};
