+++ leptos_hotkeys/src/macros.rs
@@ -110,7 +110,7 @@ macro_rules! use_hotkeys {
 macro_rules! use_hotkeys_ref {
     (($node_ref:expr, $key_combo:literal) => $($code:tt)*) => {
         {
-            $crate::use_hotkeys_ref_scoped(
+            $crate::use_hotkeys_ref(
                 $node_ref,
                 $key_combo.to_string(),
                 ::leptos::Callback::new(
@@ -123,7 +123,7 @@ macro_rules! use_hotkeys_ref {
 
     (($node_ref:expr, $key_combo:expr) => $($code:tt)*) => {
         {
-            $crate::use_hotkeys_ref_scoped(
+            $crate::use_hotkeys_ref(
                 $node_ref,
                 $key_combo.to_string(),
                 ::leptos::Callback::new(
@@ -136,7 +136,7 @@ macro_rules! use_hotkeys_ref {
 
     (($node_ref:expr, $key_combo:expr $(, $scopes:literal)*) => $($code:tt)*) => {
         {
-            $crate::use_hotkeys_ref_scoped(
+            $crate::use_hotkeys_ref(
                 $node_ref,
                 $key_combo.to_string(),
                 ::leptos::Callback::new(
@@ -149,7 +149,7 @@ macro_rules! use_hotkeys_ref {
 
     (($node_ref:expr, $key_combo:literal $(, $scopes:literal)*) => $($code:tt)*) => {
         {
-            $crate::use_hotkeys_ref_scoped(
+            $crate::use_hotkeys_ref(
                 $node_ref,
                 $key_combo.to_string(),
                 ::leptos::Callback::new(
@@ -162,7 +162,7 @@ macro_rules! use_hotkeys_ref {
 
     (($key_combo:literal $(, $scopes:expr)*) => $($code:tt)*) => {
         {
-            $crate::use_hotkeys_ref_scoped(
+            $crate::use_hotkeys_ref(
                 $key_combo.to_string(),
                 ::leptos::Callback::new(
                     $($code)*
@@ -174,7 +174,7 @@ macro_rules! use_hotkeys_ref {
 
     (($key_combo:expr $(, $scopes:expr)*) => $($code:tt)*) => {
         {
-            $crate::use_hotkeys_ref_scoped(
+            $crate::use_hotkeys_ref(
                 $key_combo.to_string(),
                 ::leptos::Callback::new(
                     $($code)*
